Rename src/hugr/{replace=>rewrite}(,.rs)
